Adding and removing nodes/edges from gexf objects

Manipulates gexf objects adding and removing nodes and edges from both, its dataframe representation and its XML representation.

add.gexf.node(
  graph,
  id = NA,
  label = NA,
  start = NULL,
  end = NULL,
  vizAtt = list(color = NULL, position = NULL, size = NULL, shape = NULL, image = NULL),
  atts = NULL
)

add.gexf.edge(
  graph,
  source,
  target,
  id = NULL,
  type = NULL,
  label = NULL,
  start = NULL,
  end = NULL,
  weight = 1,
  vizAtt = list(color = NULL, thickness = NULL, shape = NULL),
  atts = NULL,
  digits = getOption("digits")
)

rm.gexf.node(graph, id = NULL, number = NULL, rm.edges = TRUE)

rm.gexf.edge(graph, id = NULL, number = NULL)

add.node.spell(
  graph,
  id = NULL,
  number = NULL,
  start = NULL,
  end = NULL,
  digits = getOption("digits")
)

add.edge.spell(
  graph,
  id = NULL,
  number = NULL,
  start = NULL,
  end = NULL,
  digits = getOption("digits")
)

Arguments

graph
A gexf-class object.

id
A node/edge id (normally numeric value).

label
A node/edge label.

start
Starting time period

end
Ending time period

vizAtt
A list of node/edge viz attributes (see write.gexf()).

atts
List of attributes, currently ignored.

source
Source node's id.

target
Target node's id.

type
Type of connection (edge).

weight
Edge weight.

digits
Integer. Number of decimals to keep for nodes/edges sizes. See print.default()

number
Index number(s) of a single or a group of nodes or edges.

rm.edges
Whether to remove or not existing edges.

Value

A gexf object (see write.gexf()).

Details

new.gexf.graph Creates a new gexf empty object (0 nodes 0 edges).

add.gexf.node and add.gexf.edge allow adding nodes and edges to a gexf object (graph) one at a time. rm.gexf.node and rm.gexf.edges remove nodes and edges respectively.

In the case of rm.gexf.node, by default every edge linked to the node that is been removed will also be removed (rm.edges = TRUE).

add.node.spell and add.edge.spell allow to include spells to specific nodes or edges in a gexf object.

References

The GEXF project website: http://gexf.net

Examples

if (interactive()) {
  demo(gexfbuildfromscratch)
}