S3 methods for gexf objects

Methods to print and summarize gexf class objects

# S3 method for gexf
print(x, file = NA, replace = F, ...)

# S3 method for gexf
summary(object, ...)

Arguments

x
An gexf class object.

file
String. Output path where to save the GEXF file.

replace
Logical. If file exists, TRUE would replace the file.

...
Ignored

object
An gexf class object.

Value

list("print.gexf")

None (invisible NULL).

list("summary.gexf")

List containing some gexf object statistics.

Details

print.gexf displays the graph (XML) in the console. If file is not NA, a GEXF file will be exported to the indicated filepath.

summay.gexf prints summary statistics and information about the graph.

See also

Examples

  if (interactive()) {
    # Data frame of nodes
    people <- data.frame(id=1:4, label=c("juan", "pedro", "matthew", "carlos"),
                     stringsAsFactors=F)

    # Data frame of edges
    relations <- data.frame(source=c(1,1,1,2,3,4,2,4,4),
                        target=c(4,2,3,3,4,2,4,1,1))

    # Building gexf graph
    mygraph <- gexf(nodes=people, edges=relations)

    # Summary and pring
    summary(mygraph)

    write.gexf(mygraph, output="mygraph.gexf", replace=TRUE)

    # Plotting
    plot(mygraph)

  }